+++ command/agent/command.go
@@ -558,7 +558,7 @@ WAIT:
 	// Check if this is a SIGHUP
 	if sig == syscall.SIGHUP {
 		if conf := c.handleReload(config); conf != nil {
-			*config = *conf
+			config = conf.Copy()
 		}
 		goto WAIT
 	}